Added sticky flag to navigate component
Added a stycky flag that will make the dashboard always return to the
dashboard in the args after a period of inactivity. Clicking the
dashboard resets (not cancels) the timeout. This is useful if you want a
main dashboard that the user always returns to after not using the
dashboard for a while.

Setting the sticky flag to 0 or omitting it will cause the default
timeout behaviour: to cancel the timeout when the user clicks something.

Example usage

args:
timeout: 10
sticky: 1
return: main