Change how derived widget definitions are built.
Until now, the yaml files were read line by line, replacing variables ("{{foo}}") on the fly.
If those variables contained quotes, that would easily result in an unparseable yaml file.

With this change, the yaml files are first parsed as-is. Afterwards, the resulting yaml document is traversed, replacing variables in strings. This avoids the issues with quotes completely.