Update basetemperature.js
When the dashboard is freshly compiled, the current temperature bar is set at the minimum temperature, not the current temperature. This change sets the current temperature bar to current temperature if it hasn't been set yet. All this assumes that the call to self.gauge.update at the end is correct.